+++ components/tabs/tab.jsx
@@ -10,7 +10,14 @@ export default defineComponent({
     components: { CloseCircleFilled },
     props: tabProps,
     setup(props, ctx) {
-        const { valueRef, closableRef, isCard, handleTabClick, handleClose } = inject(TABS_INJECTION_KEY);
+        const {
+            valueRef,
+            closableRef,
+            isCard,
+            handleTabClick,
+            handleClose,
+            closeModeRef,
+        } = inject(TABS_INJECTION_KEY);
 
         const mergeClosable = computed(() => {
             if (!isCard.value) return;
@@ -39,11 +46,15 @@ export default defineComponent({
                     class={{
                         [`${prefixCls}-tab`]: true,
                         [`${prefixCls}-tab-card`]: isCard.value,
-                        [`${prefixCls}-tab-active`]: valueRef.value === props.value,
+                        [`${prefixCls}-tab-active`]:
+                            valueRef.value === props.value,
                         [`${prefixCls}-tab-disabled`]: props.disabled,
+                        hover: closeModeRef.value === 'hover',
                     }}
                 >
-                    <div class={`${prefixCls}-tab-label`}>{defaultSlot ? defaultSlot() : props.name}</div>
+                    <div class={`${prefixCls}-tab-label`}>
+                        {defaultSlot ? defaultSlot() : props.name}
+                    </div>
                     {mergeClosable.value && (
                         <div class={`${prefixCls}-tab-close`}>
                             <CloseCircleFilled onClick={handleCloseClick} />