+++ components/tree/useState.js
@@ -7,43 +7,37 @@ export default (props, { emit }) => {
         emit,
         { prop: 'expandedKeys', isEqual: true },
     );
+
     const [currentCheckedKeys, updateCheckedKeys] = useNormalModel(
         props,
         emit,
         { prop: 'checkedKeys', isEqual: true },
     );
+
     const [currentSelectedKeys, updateSelectedKeys] = useNormalModel(
         props,
         emit,
         { prop: 'selectedKeys', isEqual: true },
     );
+
     const { filter, hiddenKeys, filteredExpandedKeys } = useFilter(props);
 
     const hasSelected = (value) => currentSelectedKeys.value.includes(value);
     const hasChecked = (value) => currentCheckedKeys.value.includes(value);
     const hasIndeterminate = (node) => {
-        const value = node[props.valueField];
-        if (hasChecked(value)) {
+        if (hasChecked(node.value)) {
             return false;
         }
         if (node.isLeaf) {
             return false;
         }
         return (
+            props.cascade &&
             Array.isArray(node.children) &&
             node.children.some(
-                (item) =>
-                    hasChecked(item[props.valueField]) ||
-                    hasIndeterminate(item),
+                (item) => hasChecked(node.value) || hasIndeterminate(item),
             )
         );
-    };
-
-    const hasExpanded = (value) => {
-        if (filteredExpandedKeys.length) {
-            return filteredExpandedKeys.includes(value);
-        }
-        return currentExpandedKeys.value.includes(value);
     };
 
     return {
@@ -59,6 +53,5 @@ export default (props, { emit }) => {
         hasSelected,
         hasChecked,
         hasIndeterminate,
-        hasExpanded,
     };
 };