+++ components/version/index.ts
@@ -1,3 +1,3 @@
-const version = '0.1.5';
+const version = '0.1.9';
 
 export default version;