+++ docs/.vitepress/components/form/validate.vue
@@ -2,7 +2,7 @@
     <FForm ref="WFormDomRef" labelWidth="140px" labelPosition="right" :model="modelForm" :rules="rules">
         <FFormItem prop="name" :rules="nameRules">
             <template v-slot:label><span>输入姓名(slot)</span></template>
-            <FInput v-model="modelForm.name" placeholder="请输入" @change="changeHandler"></FInput>
+            <FInput v-model="modelForm.name" placeholder="请输入" @input="changeHandler"></FInput>
         </FFormItem>
         <FFormItem label="输入密码" prop="password">
             <FInput v-model="modelForm.password" type="password" showPassword placeholder="请输入密码" @input="handlePasswordInput"></FInput>
@@ -29,11 +29,12 @@
                 <FRadio value="2">女</FRadio>
             </FRadioGroup>
         </FFormItem>
-        <FFormItem label="年龄范围" prop="age">
-            <FCheckboxGroup v-model="modelForm.age" @change="changeHandler">
-                <FCheckbox value="1-10">1-10</FCheckbox>
-                <FCheckbox value="11-30">11-30</FCheckbox>
-                <FCheckbox value="31-60">31-60</FCheckbox>
+        <FFormItem label="操作权限" prop="permission">
+            <FCheckboxGroup v-model="modelForm.permission" @change="changeHandler">
+                <FCheckbox :value="1">Admin</FCheckbox>
+                <FCheckbox :value="2">edit</FCheckbox>
+                <FCheckbox :value="3">run</FCheckbox>
+                <FCheckbox :value="4">view</FCheckbox>
             </FCheckboxGroup>
         </FFormItem>
         <FFormItem prop="more" labelClass="more-label">
@@ -88,7 +89,7 @@ export default {
             time: '',
             sex: '2',
             show: true,
-            age: ['1-10'],
+            permission: [],
             more: {
                 height: 180,
                 weight: 100
@@ -111,7 +112,7 @@ export default {
 
         const rules = {
             name: [
-                { min: 3, max: 8, message: '姓名长度在 3 到 8 个字符', trigger: ['input'] },
+                { min: 3, max: 8, message: '姓名长度在 3 到 8 个字符', trigger: 'input' },
             ],
             password: [
                 { required: true, message: '请输入密码', trigger: ['blur', 'input'] }
@@ -138,8 +139,8 @@ export default {
             sex: [{
                 required: true, message: '请选择性别', trigger: 'change'
             }],
-            age: [{
-                required: true, message: '请选择年龄范围', trigger: 'change', type: 'array'
+            permission: [{
+                required: true, message: '请选择权限', trigger: 'change', type: 'array'
             }],
             desc: [
                 { required: true, message: '请备注内容', trigger: ['blur'] },
@@ -180,9 +181,9 @@ export default {
         }
 
         const submitHandler = async () => {
-            
             WFormDomRef.value.validate().then((result) => {
                 console.log('表单验证成功: ', result);
+                WFormDomRef.value.resetFields();
             }).catch((error) => {
                 console.log('表单验证失败: ', error);
             })